consolidating types and classes into Sync.ts, fixed regex id extraction
