--- src/classes/File.ts
@@ -1,8 +1,0 @@
-export default class File {
-    filename: string;
-    lines: string[];
-    constructor(filename: string) {
-      this.filename = filename;
-      this.lines = [];
-    }
-  }
